# crates/backend/openapi.yaml
openapi: "3.0.3"
info:
  title: SubReader Backend API
  description: |
    SubReader cloud sync backend service API documentation.

    Provides account management, multi-device data synchronization, and file storage capabilities.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via the login endpoint.
    Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Error Handling
    All error responses follow a unified format:
    ```json
    {
      "error": "error_type",
      "message": "Human-readable error description"
    }
    ```
  version: "0.1.0"
  contact:
    name: SubReader Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login

  # ─── Reusable Schemas ─────────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error type identifier
          example: bad_request
        message:
          type: string
          description: Human-readable error message
          example: Invalid input

    # ── Health ──
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Operation completed successfully

    # ── Devices ──
...
paths:
...
  /files/{file_id}:
    get:
      tags: [Files]
      summary: Download a file
      description: |
        Download a file by its ID.
        Returns the raw file content as `application/octet-stream` with a
        `Content-Disposition` header for the original file name.

        The response body is streamed by the server to avoid loading the
        entire file into memory. HTTP Range requests are not currently supported.
      operationId: downloadFile
      security:
        - BearerAuth: []
      parameters:
        - name: file_id
          in: path
          required: true
          description: File ID
          schema:
            type: string
            format: uuid
            example: fc73b66d-5381-43b2-a541-c8b7eb05a32e
      responses:
        "200":
          description: File content
          headers:
            Content-Disposition:
              description: 'Attachment with original filename'
              schema:
                type: string
                example: 'attachment; filename="my-book.epub"'
            Content-Length:
              description: File size in bytes
              schema:
                type: integer
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Files]
      summary: Delete a file
      description: |
        Soft-delete a file. The file metadata is marked as deleted
        but not immediately purged from storage.
      operationId: deleteFile
      security:
        - BearerAuth: []
      parameters:
        - name: file_id
          in: path
          required: true
          description: File ID to delete
          schema:
            type: string
            format: uuid
            example: fc73b66d-5381-43b2-a541-c8b7eb05a32e
      responses:
        "200":
          description: File deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: File deleted successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"